feat: 알림 목록의 항목 클릭 리스너 연결

+++ android/2023-emmsale/app/src/main/java/com/emmsale/data/notification/updated/ChildCommentNotification.kt
@@ -3,9 +3,9 @@ package com.emmsale.data.notification.updated
 import java.time.LocalDateTime
 
 class ChildCommentNotification(
-    id: Int,
-    receiverId: Int,
-    redirectId: Int,
+    id: Long,
+    receiverId: Long,
+    redirectId: Long,
     createdAt: LocalDateTime,
     isRead: Boolean,
     val commentContent: String,
