+++ android/2023-emmsale/app/src/main/java/com/emmsale/data/notification/updated/InterestEventNotification.kt
@@ -3,9 +3,9 @@ package com.emmsale.data.notification.updated
 import java.time.LocalDateTime
 
 class InterestEventNotification(
-    id: Int,
-    receiverId: Int,
-    redirectId: Int,
+    id: Long,
+    receiverId: Long,
+    redirectId: Long,
     createdAt: LocalDateTime,
     isRead: Boolean,
 ) : UpdatedNotification(
