+++ android/2023-emmsale/app/src/main/java/com/emmsale/data/notification/updated/UpdatedNotification.kt
@@ -3,9 +3,9 @@ package com.emmsale.data.notification.updated
 import java.time.LocalDateTime
 
 abstract class UpdatedNotification(
-    val id: Int,
-    val receiverId: Int,
-    val redirectId: Int,
+    val id: Long,
+    val receiverId: Long,
+    val redirectId: Long,
     val createdAt: LocalDateTime,
     val isPast: Boolean,
 )
