+++ android/2023-emmsale/app/src/main/java/com/emmsale/data/notification/updated/dto/UpdatedNotificationApiModel.kt
@@ -4,11 +4,11 @@ import kotlinx.serialization.SerialName
 
 data class UpdatedNotificationApiModel(
     @SerialName("updateNotificationId")
-    val id: Int,
+    val id: Long,
     @SerialName("receiverId")
-    val receiverId: Int,
+    val receiverId: Long,
     @SerialName("redirectId")
-    val redirectId: Int,
+    val redirectId: Long,
     @SerialName("createdAt")
     val createdAt: String, // 2023:08:23:12:00:00
     @SerialName("type")
